getting rid of the pdf version of the schedule and all the stuff needed to parse it.
