+++ content/14_team_league_schedule.txt
@@ -1,0 +1,13 @@
+1-2 3-4 5-6 7-8 9-10 11-12 13-14
+11-8 13-10 1-12 3-14 2-5 4-7 6-9
+10-5 12-7 9-14 2-11 4-13 6-1 8-3
+9-4 11-6 8-13 10-1 3-12 5-14 7-2
+14-7 2-9 4-11 6-13 1-8 10-3 12-5
+6-3 5-8 10-7 9-12 11-14 13-2 1-4
+2-12 6-14 11-3 1-7 10-4 8-9 5-13
+8-10 1-3 7-9 5-4 13-11 2-6 14-12
+4-14 10-2 11-5 13-3 6-7 12-8 9-1
+7-11 14-1 6-4 12-10 5-3 9-13 2-8
+12-13 9-5 14-8 4-2 6-10 1-11 3-7
+3-9 4-12 13-1 8-6 14-2 7-5 11-10
+5-1 7-13 3-2 11-9 12-6 14-10 4-8
